Pixaget Builder was split in to two, RetrofitBuilder and OkHttpBuilder, in order to limit the dependencies the user will have to add to his project. Those allowing the user to choose witch implementing he\she prefer.
A clear message was added to the NoClassFoundException that offers a solution if the user forgot to add the necessary dependencies.

A conflict of dependencies between the Moshi for kotlin and the Retrofit MoshiConverterFactory was fixed.

A KotlinJsonAdapterFactory was added into the MoshiAdapter.kt to solve some discrepancies and irregularity in the search results from the Pixabay API.

Move test Api key to local pc